fix home page styles
