+++ Site/webapp/wdkCustomization/js/client/components/Page/HomePage/HomePage.tsx
@@ -3,101 +3,27 @@ import { useSelector } from "react-redux";
 
 import { RootState } from "wdk-client/Core/State/Types";
 
-import { lighten } from "@material-ui/core/styles";
-import { makeStyles, createStyles, Theme } from "@material-ui/core";
-
 import Grid from "@material-ui/core/Grid";
 import Box from "@material-ui/core/Box";
 
 import { SearchPanel, AvailableDataPanel, StatsPanel, AboutPanel } from "./Panels";
+import useHomePageStyles from "./styles"
 
 import "./HomePage.scss";
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        primaryBackground: {
-            background: theme.palette.primary.main,
-        },
-        lightBackground: {
-            background: lighten(theme.palette.primary.main, 0.95),
-        },
-        defaultBackgroundPanel: {
-            paddingTop: theme.spacing(6),
-        },
-        darkContrastText: {
-            color: theme.palette.primary.contrastText,
-        },
-        lightContrastText: {
-            color: theme.palette.primary.light,
-        },
-        fancyBody: {
-            fontSize: "1.7rem",
-             fontFamily: '"Raleway", "Roboto", "Arial", "sans-serif"'
-        },
-        largeBody: {
-            fontSize: "1.2rem"
-        },
-        secondaryText: {
-            color: theme.palette.secondary.main,
-        },
-        bold: {
-            fontWeight: theme.typography.fontWeightBold,
-        },
-        primaryText: {
-            color: theme.palette.primary.main,
-        },
-        largeIcon: {
-            fontSize: 65,
-        },
-        largeButton: {
-            fontSize: "1.1rem",
-        },
-        noTopPadding: {
-            paddingTop: "0px",
-        },
-        extraTopPadding: {
-            paddingTop: theme.spacing(6),
-        },
-        highlightStat: {
-            fontSize: "2rem",
-        },
-        donutChart: {
-            maxWidth: 900,
-            /*[theme.breakpoints.down("sm")]: {
-                maxHeight: 300,
-                maxWidth: 500,
-            },*/
-        },
-        smallCaps: {
-            fontVariant: "all-small-caps",
-        },
-        darkBgLink: {
-            color: theme.palette.secondary.main,
-            "&:hover": {
-                color: theme.palette.secondary.light,
-            },
-        },
-        lightBgLink: {
-            color: theme.palette.secondary.dark,
-            "&:hover": {
-                color: theme.palette.secondary.light,
-            },
-        },
-    })
-);
 
 export const HomePage: React.FC<any> = ({}) => {
     //const endpoint = useSelector((state: RootState) => state.globalData?.siteConfig?.endpoint);
     //const projectId = useSelector((state: RootState) => state.globalData?.config?.projectId);
     const webAppUrl = useSelector((state: RootState) => state.globalData?.siteConfig?.webAppUrl);
-    const classes = useStyles();
+    const classes = useHomePageStyles();
 
     return (
         <Grid container direction="column" alignItems="center">
-            <SearchPanel classes={classes} />
-            <AvailableDataPanel classes={classes} webAppUrl={webAppUrl} />
-            <StatsPanel classes={classes} />
-            <AboutPanel classes={classes} webAppUrl={webAppUrl} />
+            <SearchPanel />
+            <AvailableDataPanel  webAppUrl={webAppUrl} />
+            <StatsPanel  />
+            <AboutPanel webAppUrl={webAppUrl} />
             <Box pb={4}/>
         </Grid>
     );
