+++ Site/webapp/wdkCustomization/js/client/components/Page/HomePage/Panels/AboutPanel.tsx
@@ -8,10 +8,12 @@ import { PanelProps, DefaultBackgroundPanel } from "@components/MaterialUI";
 import { _externalUrls } from "genomics-client/data/_externalUrls";
 import { _siteStatistics } from "genomics-client/data/_siteStatistics";
 
+import useHomePageStyles from "../styles";
 
-export const AboutPanel: React.FC<PanelProps> = ({ classes, background = "light", webAppUrl }) => {
+export const AboutPanel: React.FC<PanelProps> = ({ background = "light", webAppUrl }) => {
+    const classes = useHomePageStyles();
     const bodyTextColor = background === "dark" ? classes.darkContrastText : classes.lightContrastText;
-    const headingTextColor = background === "dark" ? classes.headingSecondary : classes.headingPrimary;
+    const headingTextColor = background === "dark" ? classes.secondaryText : classes.primaryText;
     const linkType = background === "dark" ? classes.darkBgLink : classes.lightBgLink;
     const bodyText = bodyTextColor + " " + classes.largeBody;
