--- Site/webapp/wdkCustomization/js/client/components/Record/RecordHeading/ImpactIndicator/ImpactIndicator.scss
@@ -1,17 +1,0 @@
-.impact-indicator{
-	&.high {
-		color: $high;
-	}
-	&.moderate {
-		color: $moderate;
-	}
-	&.low{
-		color: $low;
-	}
-	&.modifier {
-		color: $modifier;
-	}
-	&.notreported {
-		color: $notReported;
-	}
-}
