+++ Site/webapp/wdkCustomization/sass/client.scss
@@ -26,4 +26,4 @@ $theme-colors: (
 @import "../js/client/components/Components";
 @import "./wdk";
 @import "./style";
-
+@import "./variant-classification";
